DOMCore/tests/submissions/Ms2ger/Node-removeChild.html

<!DOCTYPE html>
<title>Node.removeChild</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src=about:blank></iframe>
<script>
test(function() {
  var frameDoc = frames[0].document,
      s = frameDoc.createElement("a")
  assert_equals(s.ownerDocument, frameDoc)
  assert_throws("NOT_FOUND_ERR", function() { document.body.removeChild(s) })
  assert_equals(s.ownerDocument, frameDoc)

  s = frameDoc.createElement("b")
  frameDoc.body.appendChild(s)
  assert_equals(s.ownerDocument, frameDoc)
  assert_throws("NOT_FOUND_ERR", function() { document.body.removeChild(s) })
  assert_equals(s.ownerDocument, frameDoc)

  assert_throws(new TypeError(), function() { document.body.removeChild(null) })

  assert_throws(new TypeError(), function() { document.body.removeChild({'a':'b'}) })

  document.body.setAttribute("abc", "test")
  var attr = document.body.attributes[0]
  assert_throws("NOT_SUPPORTED_ERR", function() { attr.removeChild(null) })
  assert_equals(attr.value, "test")
  assert_equals(attr.childNodes.length, 1)
  assert_equals(attr.childNodes[0].nodeValue, "test")
  attr.removeChild(attr.firstChild)
  assert_equals(document.body.getAttribute("abc"), "")
  assert_true(document.body.hasAttribute("abc"))
  assert_equals(attr.value, "")
  assert_equals(attr.childNodes.length, 0)
})
</script>
